Added docs + features to TwiTransfer class
Added a much more verbose documentation for the initialize method
of the TwiTransfer class. Debugged some issues with OAuth and got
the gem to work nicely with the application. Refactored
get_access_token to make it a little more user friendly and elegant.

Added .access_request to make building the request path a little more
modular and less garrish. Still more work to do on tests, but its
coming along nicely. Next stop is taking the response and forming
the POST request to the target user's friends list.